chore: Removed dependency on cd/cv for reconfiguring operation
