+++ controllers/apps/operations/reconfigure_util.go
@@ -20,21 +20,15 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 package operations
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 
 	"github.com/spf13/cast"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	appsv1alpha1 "github.com/apecloud/kubeblocks/apis/apps/v1alpha1"
 	"github.com/apecloud/kubeblocks/pkg/configuration/core"
-	"github.com/apecloud/kubeblocks/pkg/constant"
-	"github.com/apecloud/kubeblocks/pkg/controller/configuration"
-	intctrlutil "github.com/apecloud/kubeblocks/pkg/controllerutil"
 )
 
 type reconfiguringResult struct {
@@ -45,113 +39,113 @@ type reconfiguringResult struct {
 	err                  error
 }
 
-type updateReconfigureStatus func(params []core.ParamPairs, orinalData map[string]string, formatter *appsv1alpha1.FormatterConfig) error
+// type updateReconfigureStatus func(params []core.ParamPairs, orinalData map[string]string, formatter *appsv1alpha1.FormatterConfig) error
 
 // Deprecated: use NewPipeline instead
 // updateConfigConfigmapResource merges parameters of the config into the configmap, and verifies final configuration file.
-func updateConfigConfigmapResource(config appsv1alpha1.ConfigurationItem,
-	configSpec appsv1alpha1.ComponentConfigSpec,
-	cmKey client.ObjectKey,
-	ctx context.Context,
-	cli client.Client,
-	opsCrName string,
-	updater updateReconfigureStatus) reconfiguringResult {
-	var (
-		cm = &corev1.ConfigMap{}
-		cc = &appsv1alpha1.ConfigConstraint{}
+// func updateConfigConfigmapResource(config appsv1alpha1.ConfigurationItem,
+//	configSpec appsv1alpha1.ComponentConfigSpec,
+//	cmKey client.ObjectKey,
+//	ctx context.Context,
+//	cli client.Client,
+//	opsCrName string,
+//	updater updateReconfigureStatus) reconfiguringResult {
+//	var (
+//		cm = &corev1.ConfigMap{}
+//		cc = &appsv1alpha1.ConfigConstraint{}
+//
+//		err    error
+//		newCfg map[string]string
+//	)
+//
+//	if err := cli.Get(ctx, cmKey, cm); err != nil {
+//		return makeReconfiguringResult(err)
+//	}
+//	if err := cli.Get(ctx, client.ObjectKey{
+//		Namespace: configSpec.Namespace,
+//		Name:      configSpec.ConfigConstraintRef,
+//	}, cc); err != nil {
+//		return makeReconfiguringResult(err)
+//	}
+//
+//	updatedFiles := make(map[string]string, len(config.Keys))
+//	updatedParams := make([]core.ParamPairs, 0, len(config.Keys))
+//	for _, key := range config.Keys {
+//		if key.FileContent != "" {
+//			updatedFiles[key.Key] = key.FileContent
+//		}
+//		if len(key.Parameters) > 0 {
+//			updatedParams = append(updatedParams, core.ParamPairs{
+//				Key:           key.Key,
+//				UpdatedParams: fromKeyValuePair(key.Parameters),
+//			})
+//		}
+//	}
+//
+//	if newCfg, err = mergeUpdatedParams(cm.Data, updatedFiles, updatedParams, cc, configSpec); err != nil {
+//		return makeReconfiguringResult(err, withFailed(true))
+//	}
+//	configPatch, restart, err := core.CreateConfigPatch(cm.Data, newCfg, cc.Spec.FormatterConfig.Format, configSpec.Keys, len(updatedFiles) != 0)
+//	if err != nil {
+//		return makeReconfiguringResult(err)
+//	}
+//	if !restart && !configPatch.IsModify {
+//		return makeReconfiguringResult(nil, withReturned(newCfg, configPatch))
+//	}
+//	if updater != nil {
+//		if err := updater(updatedParams, cm.Data, cc.Spec.FormatterConfig); err != nil {
+//			return makeReconfiguringResult(err)
+//		}
+//	}
+//
+//	return makeReconfiguringResult(
+//		syncConfigmap(cm, newCfg, cli, ctx, opsCrName, configSpec, &cc.Spec, config.Policy),
+//		withReturned(newCfg, configPatch),
+//		withNoFormatFilesUpdated(restart))
+// }
 
-		err    error
-		newCfg map[string]string
-	)
+// func mergeUpdatedParams(base map[string]string,
+//	updatedFiles map[string]string,
+//	updatedParams []core.ParamPairs,
+//	cc *appsv1alpha1.ConfigConstraint,
+//	tpl appsv1alpha1.ComponentConfigSpec) (map[string]string, error) {
+//	updatedConfig := base
+//
+//	// merge updated files into configmap
+//	if len(updatedFiles) != 0 {
+//		return core.MergeUpdatedConfig(base, updatedFiles), nil
+//	}
+//	if cc == nil {
+//		return updatedConfig, nil
+//	}
+//	return intctrlutil.MergeAndValidateConfigs(cc.Spec, updatedConfig, tpl.Keys, updatedParams)
+// }
 
-	if err := cli.Get(ctx, cmKey, cm); err != nil {
-		return makeReconfiguringResult(err)
-	}
-	if err := cli.Get(ctx, client.ObjectKey{
-		Namespace: configSpec.Namespace,
-		Name:      configSpec.ConfigConstraintRef,
-	}, cc); err != nil {
-		return makeReconfiguringResult(err)
-	}
-
-	updatedFiles := make(map[string]string, len(config.Keys))
-	updatedParams := make([]core.ParamPairs, 0, len(config.Keys))
-	for _, key := range config.Keys {
-		if key.FileContent != "" {
-			updatedFiles[key.Key] = key.FileContent
-		}
-		if len(key.Parameters) > 0 {
-			updatedParams = append(updatedParams, core.ParamPairs{
-				Key:           key.Key,
-				UpdatedParams: fromKeyValuePair(key.Parameters),
-			})
-		}
-	}
-
-	if newCfg, err = mergeUpdatedParams(cm.Data, updatedFiles, updatedParams, cc, configSpec); err != nil {
-		return makeReconfiguringResult(err, withFailed(true))
-	}
-	configPatch, restart, err := core.CreateConfigPatch(cm.Data, newCfg, cc.Spec.FormatterConfig.Format, configSpec.Keys, len(updatedFiles) != 0)
-	if err != nil {
-		return makeReconfiguringResult(err)
-	}
-	if !restart && !configPatch.IsModify {
-		return makeReconfiguringResult(nil, withReturned(newCfg, configPatch))
-	}
-	if updater != nil {
-		if err := updater(updatedParams, cm.Data, cc.Spec.FormatterConfig); err != nil {
-			return makeReconfiguringResult(err)
-		}
-	}
-
-	return makeReconfiguringResult(
-		syncConfigmap(cm, newCfg, cli, ctx, opsCrName, configSpec, &cc.Spec, config.Policy),
-		withReturned(newCfg, configPatch),
-		withNoFormatFilesUpdated(restart))
-}
-
-func mergeUpdatedParams(base map[string]string,
-	updatedFiles map[string]string,
-	updatedParams []core.ParamPairs,
-	cc *appsv1alpha1.ConfigConstraint,
-	tpl appsv1alpha1.ComponentConfigSpec) (map[string]string, error) {
-	updatedConfig := base
-
-	// merge updated files into configmap
-	if len(updatedFiles) != 0 {
-		return core.MergeUpdatedConfig(base, updatedFiles), nil
-	}
-	if cc == nil {
-		return updatedConfig, nil
-	}
-	return intctrlutil.MergeAndValidateConfigs(cc.Spec, updatedConfig, tpl.Keys, updatedParams)
-}
-
-func syncConfigmap(
-	cmObj *corev1.ConfigMap,
-	newCfg map[string]string,
-	cli client.Client,
-	ctx context.Context,
-	opsCrName string,
-	configSpec appsv1alpha1.ComponentConfigSpec,
-	cc *appsv1alpha1.ConfigConstraintSpec,
-	policy *appsv1alpha1.UpgradePolicy) error {
-
-	patch := client.MergeFrom(cmObj.DeepCopy())
-	cmObj.Data = newCfg
-	if cmObj.Annotations == nil {
-		cmObj.Annotations = make(map[string]string)
-	}
-	if policy != nil {
-		cmObj.Annotations[constant.UpgradePolicyAnnotationKey] = string(*policy)
-	}
-	cmObj.Annotations[constant.LastAppliedOpsCRAnnotationKey] = opsCrName
-	core.SetParametersUpdateSource(cmObj, constant.ReconfigureUserSource)
-	if err := configuration.SyncEnvConfigmap(configSpec, cmObj, cc, cli, ctx); err != nil {
-		return err
-	}
-	return cli.Patch(ctx, cmObj, patch)
-}
+// func syncConfigmap(
+//	cmObj *corev1.ConfigMap,
+//	newCfg map[string]string,
+//	cli client.Client,
+//	ctx context.Context,
+//	opsCrName string,
+//	configSpec appsv1alpha1.ComponentConfigSpec,
+//	cc *appsv1alpha1.ConfigConstraintSpec,
+//	policy *appsv1alpha1.UpgradePolicy) error {
+//
+//	patch := client.MergeFrom(cmObj.DeepCopy())
+//	cmObj.Data = newCfg
+//	if cmObj.Annotations == nil {
+//		cmObj.Annotations = make(map[string]string)
+//	}
+//	if policy != nil {
+//		cmObj.Annotations[constant.UpgradePolicyAnnotationKey] = string(*policy)
+//	}
+//	cmObj.Annotations[constant.LastAppliedOpsCRAnnotationKey] = opsCrName
+//	core.SetParametersUpdateSource(cmObj, constant.ReconfigureUserSource)
+//	if err := configuration.SyncEnvConfigmap(configSpec, cmObj, cc, cli, ctx); err != nil {
+//		return err
+//	}
+//	return cli.Patch(ctx, cmObj, patch)
+// }
 
 func updateOpsLabelWithReconfigure(obj *appsv1alpha1.OpsRequest, params []core.ParamPairs, orinalData map[string]string, formatter *appsv1alpha1.FormatterConfig) {
 	var maxLabelCount = 16
@@ -250,14 +244,6 @@ func makeReconfiguringResult(err error, ops ...func(*reconfiguringResult)) recon
 	return result
 }
 
-func getConfigSpecName(configSpec []appsv1alpha1.ComponentConfigSpec) []string {
-	names := make([]string, len(configSpec))
-	for i, spec := range configSpec {
-		names[i] = spec.Name
-	}
-	return names
-}
-
 func constructReconfiguringConditions(result reconfiguringResult, resource *OpsResource, configSpec *appsv1alpha1.ComponentConfigSpec) *metav1.Condition {
 	if result.noFormatFilesUpdated || result.configPatch.IsModify {
 		return appsv1alpha1.NewReconfigureRunningCondition(
@@ -320,4 +306,58 @@ func formatConfigPatchToMessage(configPatch *core.ConfigPatchInfo, execStatus *c
 		configPatch.UpdateConfig,
 		configPatch.AddConfig,
 		configPatch.DeleteConfig)
+}
+
+func updateFileContent(item *appsv1alpha1.ConfigurationItemDetail, key string, content string) {
+	params, ok := item.ConfigFileParams[key]
+	if !ok {
+		item.ConfigFileParams[key] = appsv1alpha1.ConfigParams{
+			Content: &content,
+		}
+		return
+	}
+	item.ConfigFileParams[key] = appsv1alpha1.ConfigParams{
+		Parameters: params.Parameters,
+		Content:    &content,
+	}
+}
+
+func updateParameters(item *appsv1alpha1.ConfigurationItemDetail, key string, parameters []appsv1alpha1.ParameterPair) {
+	updatedParams := make(map[string]*string, len(parameters))
+	for _, parameter := range parameters {
+		updatedParams[parameter.Key] = parameter.Value
+	}
+
+	params, ok := item.ConfigFileParams[key]
+	if !ok {
+		item.ConfigFileParams[key] = appsv1alpha1.ConfigParams{
+			Parameters: updatedParams,
+		}
+		return
+	}
+
+	item.ConfigFileParams[key] = appsv1alpha1.ConfigParams{
+		Content:    params.Content,
+		Parameters: mergeMaps(params.Parameters, updatedParams),
+	}
+}
+
+func mergeMaps(m1 map[string]*string, m2 map[string]*string) map[string]*string {
+	merged := make(map[string]*string)
+	for key, value := range m1 {
+		merged[key] = value
+	}
+	for key, value := range m2 {
+		merged[key] = value
+	}
+	return merged
+}
+
+func hasFileUpdate(config appsv1alpha1.ConfigurationItem) bool {
+	for _, key := range config.Keys {
+		if key.FileContent != "" {
+			return true
+		}
+	}
+	return false
 }
